Properly implement front comments
Instead of blindly slicing at the first colon and then checking whether the comment part is alphanumeric, this commit moves the "stray character in program" error to be a reading error rather than a lexing error. This means that there can be quotes and forms and other unaccepted characters in a front comment.