Allow printing multiple lex/parse errors
This commit changes the Hexit program type to be a vector of lines, rather than a string containing the whole program including any newlines. The change means that the `hexit_lang` reader can now produce one error per line, rather than one error maximum for the entire program that's being read.

Keeping the lines around, rather than keeping the string around, means that in the future we can show the line to the user as part of the error, the same way that the Rust compiler does.